added password encryption and users post works

+++ src/main/java/com/example/venusblog/controller/PostsController.java
@@ -3,11 +3,16 @@ package com.example.venusblog.controller;
 import com.example.venusblog.data.Category;
 import com.example.venusblog.data.Post;
 import com.example.venusblog.data.User;
+import com.example.venusblog.misc.FieldHelper;
 import com.example.venusblog.repository.CategoriesRepository;
 import com.example.venusblog.repository.PostsRepository;
 import com.example.venusblog.repository.UsersRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -36,15 +41,16 @@ public class PostsController {
     }
 
     @PostMapping("")
-    public void createPost(@RequestBody Post newPost) {
-
-        // use a fake author for the post
-        User author = usersRepository.findById(1L).get();
+    public void createPost(@RequestBody Post newPost, OAuth2Authentication auth) {
+    String  userName = auth.getName();
+    User author = usersRepository.findByUserName((userName));
+//         use a fake author for the post
+//        User author = usersRepository.findById(1L).get();
         newPost.setAuthor(author);
-
+        newPost.setCategories(new ArrayList<>());
         Category cat1 = categoriesRepository.findById(1L).get();
         Category cat2 = categoriesRepository.findById(2L).get();
-        newPost.setCategories(new ArrayList<>());
+
         newPost.getCategories().add(cat1);
         newPost.getCategories().add(cat2);
 //
@@ -62,33 +68,28 @@ public class PostsController {
 
     @DeleteMapping("/{id}")
     public void deletePostById(@PathVariable long id) {
+        Optional<Post> optionalPost = postsRepository.findById(id);
+        if(optionalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post id " + id + " not found");
+        }
         postsRepository.deleteById(id);
-        // what to do if we don't find it
-//        throw new RuntimeException("Post not found");
     }
 
     @PutMapping("/{id}")
     public void updatePost(@RequestBody Post updatedPost, @PathVariable long id) {
+        Optional<Post> originalPost = postsRepository.findById(id);
+        if(originalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post " + id + " not found");
+        }
+
         // in case id is not in the request body (i.e., updatedPost), set it
         // with the path variable id
         updatedPost.setId(id);
-        postsRepository.save(updatedPost);
-//------------------------------------------------------------------------------------------------------------
-//        // find the post to update in the posts list
-//
-//        Post post = findPostById(id);
-//        if(post == null) {
-//            System.out.println("Post not found");
-//        } else {
-//            if(updatedPost.getTitle() != null) {
-//                post.setTitle(updatedPost.getTitle());
-//            }
-//            if(updatedPost.getContent() != null) {
-//                post.setContent(updatedPost.getContent());
-//            }
-//            return;
-//        }
-//        throw new RuntimeException("Post not found");
+
+        // copy any new field values FROM updatedPost TO originalPost
+        BeanUtils.copyProperties(updatedPost, originalPost.get(), FieldHelper.getNullPropertyNames(updatedPost));
+
+        postsRepository.save(originalPost.get());
     }
