+++ src/main/java/com/example/venusblog/controller/UsersController.java
@@ -2,13 +2,18 @@ package com.example.venusblog.controller;
 
 import com.example.venusblog.data.User;
 import com.example.venusblog.data.UserRole;
+import com.example.venusblog.dto.UserFetchDTO;
+import com.example.venusblog.misc.FieldHelper;
 import com.example.venusblog.repository.UsersRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.BeanUtils;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -17,18 +22,31 @@ import java.util.Optional;
 @RequestMapping(value = "/api/users", produces = "application/json")
 public class UsersController {
     private UsersRepository usersRepository;
-
+    private PasswordEncoder passwordEncoder;
     @GetMapping("")
-    public List<User> fetchUsers() {
-        return usersRepository.findAll();
+    public List<UserFetchDTO> fetchUsers() {
+//        return usersRepository.fetchUserDTOs();
+        List<User> users = usersRepository.findAll();
+        List<UserFetchDTO> userDTOs = new ArrayList<>();
+
+        for(User user : users) {
+            UserFetchDTO userDTO = new UserFetchDTO();
+            userDTO.setId(user.getId());
+            userDTO.setUserName(user.getUserName());
+            userDTO.setEmail(user.getEmail());
+            userDTOs.add(userDTO);
+        }
+
+        return userDTOs;
     }
 
     @GetMapping("/{id}")
     public Optional<User> fetchUserById(@PathVariable long id) {
-        Optional<User> user = usersRepository.findById(id);
-        if (user.isEmpty()) {
-    throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User" + id + "not found");
-        } return user;
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
+        return optionalUser;
     }
 
     @GetMapping("/me")
@@ -53,8 +71,12 @@ public class UsersController {
 
     @PostMapping("/create")
     public void createUser(@RequestBody User newUser) {
+        // TODO: validate new user fields
         newUser.setRole(UserRole.USER);
 
+        String plainTextPassword = newUser.getPassword();
+        String encryptedPassword = passwordEncoder.encode(plainTextPassword);
+        newUser.setPassword(encryptedPassword);
 
         // don't need the below line at this point but just for kicks
         newUser.setCreatedAt(LocalDate.now());
@@ -63,23 +85,39 @@ public class UsersController {
 
     @DeleteMapping("/{id}")
     public void deleteUserById(@PathVariable long id) {
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
         usersRepository.deleteById(id);
     }
 
     @PutMapping("/{id}")
     public void updateUser(@RequestBody User updatedUser, @PathVariable long id) {
-        // find the post to update in the posts list
-//        Optional<User>= ------------------------------------------
-        updatedUser.setId(id);
-        usersRepository.save(updatedUser);
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
+        // get the user from the optional so we no longer have to deal with the optional
+        User originalUser = optionalUser.get();
+
+        // merge the changed data in updatedUser with originalUser
+        BeanUtils.copyProperties(updatedUser, originalUser, FieldHelper.getNullPropertyNames(updatedUser));
+
+        // originalUser now has the merged data (changes + original data)
+        originalUser.setId(id);
+
+        usersRepository.save(originalUser);
     }
 
     @PutMapping("/{id}/updatePassword")
     private void updatePassword(@PathVariable Long id, @RequestParam(required = false) String oldPassword, @RequestParam String newPassword) {
-        User user = usersRepository.findById(id).get();
-//        if(user == null) {
-//            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User id " + id + " not found");
-//        }
+        Optional<User> optionalUser = usersRepository.findById(id);
+        if(optionalUser.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User " + id + " not found");
+        }
+
+        User user = optionalUser.get();
 
         // compare old password with saved pw
         if(!user.getPassword().equals(oldPassword)) {
